frontend/tasks: make mock server a service
Split it to a specific file, and rename the class. Introduces a circular
dependency that we intend to address soon.
